% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{dist_parallel}
\alias{dist_parallel}
\title{Calculate distances between pairs of time series in a list.}
\usage{
dist_parallel(
  tsList,
  measureFunc = tsdiss_euclidean,
  isSymetric = TRUE,
  error_value = NaN,
  warn_error = TRUE,
  num_cores = 1,
  ...
)
}
\arguments{
\item{tsList}{List of time series (arrays).}

\item{measureFunc}{Function to be applied to all combinations
of time series. This function should have at least two parameters
for each time series. Ex: function(ts1, ts2){ cor(ts1, ts2) }}

\item{isSymetric}{Boolean. If the distance function is symmetric.}

\item{error_value}{The value returned if an error occur when calculating a
the distance for a pair of time series.}

\item{warn_error}{Boolean. If TRUE (default), a warning will rise when an
error occur during the calculations.}

\item{num_cores}{Numeric. Number of cores}

\item{...}{Additional parameters for measureFunc}

\item{isDist}{Boolean. If measureFunc is a distance function. iF TRUE,
a zero value returned by measureFunc(ts1,ts2) means that the pair
of time series ts1 and ts2 are perfectly equal.}
}
\value{
A distance or similarity matrix M whose position M_{ij}
    corresponds to distance or similarity value between time series
    i and j.
}
\description{
This function calculates the distance between all combinations of
time series in the list and returns a distance matrix. This function
is usually the first try and might work if the number of time series
and their length are not too big.
}
