% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{dist_percentile}
\alias{dist_percentile}
\title{Returns the distance value that corresponds to the desired percentile. This function
is useful when the user wants to generate networks with different distance functions
but with the same link density.}
\usage{
dist_percentile(D, percentile = 0.1, is_D_symetric = TRUE)
}
\arguments{
\item{D}{distance matrix}

\item{percentile}{(Float) The desired percentile of lower distances.}

\item{is_D_symetric}{(Boolean)}
}
\value{
Distance percentile value.
}
\description{
Returns the distance value that corresponds to the desired percentile. This function
is useful when the user wants to generate networks with different distance functions
but with the same link density.
}
