% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_to_net.R
\name{net_enn_approx}
\alias{net_enn_approx}
\title{Construct an approximated epsilon neighbor network (faster, but
approximated) from a distance matrix. Some actual nearest neighbors
may be omitted.}
\usage{
net_enn_approx(D, eps, ...)
}
\arguments{
\item{D}{Distance matrix}

\item{eps}{(Integer) k nearest-nearest neighbors where each time seires
will be connected to}

\item{...}{Other parameters to frNN() function from dbscan package.}
}
\value{
Approximated epsilon nearest-neighbor network
}
\description{
Construct an approximated epsilon neighbor network (faster, but
approximated) from a distance matrix. Some actual nearest neighbors
may be omitted.
}
