% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_to_net.R
\name{net_epsNN_approx}
\alias{net_epsNN_approx}
\title{Construct an approximated knn-network (faster, but approximated) from
a distance matrix.}
\usage{
net_epsNN_approx(D, k, ...)
}
\arguments{
\item{D}{Distance matrix}

\item{k}{(Integer) k nearest-nearest neighbors where each time seires
will be connected to}

\item{...}{Other parameters to kNN() function from dbscan package.}
}
\value{

}
\description{
Construct an approximated knn-network (faster, but approximated) from
a distance matrix.
}
