% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_net.R
\name{net_full_weighted}
\alias{net_full_weighted}
\title{Creates a weighted fully connected network.}
\usage{
net_full_weighted(D, invert_dist_as_weight = TRUE)
}
\arguments{
\item{D}{Distance matrix}

\item{invert_dist_as_weight}{Boolean, if weighted == TRUE, then the weights
become 1 - distance. This is the default behavior since most network
measures interpret higher weights as stronger connection. If TRUE,
the distance matrix D should contain only value in the interval [0,1].}
}
\value{
Fully connected network
}
\description{
Creates a weighted fully connected network.
}
