% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_to_net.R
\name{net_knn}
\alias{net_knn}
\title{Construct a knn-network from a distance matrix.}
\usage{
net_knn(D, k, num_cores = 1)
}
\arguments{
\item{D}{Distance matrix}

\item{k}{(Integer) k nearest-nearest neighbors where each time series
will be connected to}

\item{num_cores}{(Integer) Number of cores to use.}
}
\value{
k nearest-neighbor network
}
\description{
Construct a knn-network from a distance matrix.
}
