% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_to_net.R
\name{net_knn_create}
\alias{net_knn_create}
\title{Construct a knn-network from a distance matrix.}
\usage{
net_knn_create(D, k, num_cores = 1)
}
\arguments{
\item{D}{Distance matrix}

\item{k}{(Integer) k nearest-nearest neighbors where each time seires
will be connected to}

\item{num_cores}{(Integer) Number of cores to use.}
}
\value{

}
\description{
Construct a knn-network from a distance matrix.
}
