% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_to_net.R
\name{net_knn_create_approx}
\alias{net_knn_create_approx}
\title{Construct a knn-network (faster, but approximated) from a distance matrix.}
\usage{
net_knn_create_approx(D, k, ...)
}
\arguments{
\item{D}{Distance matrix}

\item{k}{(Integer) k nearest-nearest neighbors where each time seires
will be connected to}

\item{...}{Other parameters to kNN() function}
}
\value{

}
\description{
Construct a knn-network (faster, but approximated) from a distance matrix.
}
