% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_to_net.R
\name{net_significant_links}
\alias{net_significant_links}
\title{Construct a network with significant links.}
\usage{
net_significant_links(D, directed = FALSE, add_col_rownames = TRUE)
}
\arguments{
\item{D}{Distance matrix}

\item{directed}{Boolean, directed or undirected (default) network.}

\item{add_col_rownames}{Boolean. If TRUE (default), it uses the column and row
names from dist matrix as node labels.}
}
\value{

}
\description{
Some time series distance functions in ts2net return 0 when the the two
time series are statistically similar or 1 otherwise. This function is
a wrapper for the [net_enn()] function that construct networks from a
binary distance matrix (0 means statistical similar).
}
