% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_ets}
\alias{random_ets}
\title{Random event time series generator}
\usage{
random_ets(ts_length, num_events, return_marked_times = FALSE)
}
\arguments{
\item{ts_length}{Time series Length}

\item{num_events}{The number of events}

\item{return_marked_times}{Return the time indices (marked points) where
the events occur.}
}
\value{
An event (binary, 1: event, 0 otherwise) time series
}
\description{
It generates an event time series with length ts_length with
num_events events considering a uniform probability distribution.
}
