% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ts_to_windows}
\alias{ts_to_windows}
\title{Extract time windows from a time series}
\usage{
ts_to_windows(x, width, by = 1)
}
\arguments{
\item{x}{time series}

\item{width}{window length}

\item{by}{Window step. This is the number of values in and out during
the window rollover process.}
}
\value{
List of windows
}
\description{
This function is useful when constructing a network from a single
time series. The returned list can be directly used to calculate
the distance matrix D with ts_dist().
}
