% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_cor_abs}
\alias{tsdist_cor_abs}
\title{Correlation distance.}
\usage{
tsdist_cor_abs(ts1, ts2)
}
\arguments{
\item{ts1}{Array. Time series 1}

\item{ts2}{Array. Time series 2}
}
\value{
Real value [0,1] where 0 means perfect positive or negative correlation
and 1 no  correlation.
}
\description{
Calculates 1 - abs(cor(ts1, ts2)). Different from tsdist_cor, this distance
considers both strong positive and negative correlations. Zero means no
correlation.
}
