% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_cor_test}
\alias{tsdist_cor_test}
\title{Absolute, positive, or negative correlation distance with significance test.}
\usage{
tsdist_cor_test(ts1, ts2, cor_type = "abs", sig_level = 0.01, ...)
}
\arguments{
\item{ts1}{Array. Time series 1}

\item{ts2}{Array. Time series 2}

\item{cor_type}{String. "abs" (default), "+", or "-". "abs" considers the
correlation absolute value. "+" only positive correlations and "-" only
negative correlations.}

\item{sig_level}{The significance level to test if correlation is significant.
See cor.test().}

\item{...}{Additional parameters to cor.test() function.}
}
\value{
Zero iff significant, or one otherwise.
}
\description{
This function is similar to tsdist_cor(), but also performs a significance
test to check if the absolute, positive, or negative correlation distance
is significant. See cor.test() for more details. This function returns only
zero (if significant) or one.
}
