% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_dtw}
\alias{tsdist_dtw}
\title{Dynamic Time Warping (DTW) distance.}
\usage{
tsdist_dtw(ts1, ts2)
}
\arguments{
\item{ts1}{Array. Time series 1}

\item{ts2}{Array. Time series 2}
}
\value{
DTW distance
}
\description{
This function is a wrapper for the dtw() function from the dtw package.
}
