% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_mic}
\alias{tsdist_mic}
\title{Maximal information coefficient (MIC) distance.}
\usage{
tsdist_mic(ts1, ts2)
}
\arguments{
\item{ts1}{Array. Time series 1}

\item{ts2}{Array. Time series 2}
}
\value{
Distance
}
\description{
This function transforms the MIC function (from minerva package) into
a distance function.
}
