% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_parts_merge}
\alias{tsdist_parts_merge}
\title{Merge parts of distances stored in data frames.}
\usage{
tsdist_parts_merge(list_dfs, num_elements)
}
\arguments{
\item{list_dfs}{A list of data frames. Each data frame should have three
columns i,j, and dist.}

\item{num_elements}{The number of time series in the data set. The number of elements
defines the number of rows ans columns in the distance matrix D.}
}
\value{
Distance matrix D
}
\description{
The functions tsdist_dir_parallel and tsdist_parts_parallel calculate part of
the distance matrix D. This function merges these files and construct
a distance matrix D.
}
