% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_voi}
\alias{tsdist_voi}
\title{Variation of Information distance}
\usage{
tsdist_voi(ts1, ts2, nbins = "freedman-diaconis", mi_method = "emp")
}
\arguments{
\item{ts1}{Array. Time series 1}

\item{ts2}{Array. Time series 2}

\item{nbins}{The number of bins used for the discretization of both time series.
It can be a positive integer or a string with one of the three rules
"freedman-diaconis" (default), "sturges", or "scott".}

\item{mi_method}{The name of the entropy estimator used in the functions
mutinformation() and entropy() from the infotheo package.}
}
\value{
Distance
}
\description{
The variation of information (VoI) is a distance function based on mutual
information.
}
