% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_vr_test}
\alias{tsdist_vr_test}
\title{van Rossum distance with significance test}
\usage{
tsdist_vr_test(ets1, ets2, tau = 1, repetitions = 1000, sig_level = 0.01)
}
\arguments{
\item{ets1}{Event time series 1 (one means an event, or zero otherwise)}

\item{ets2}{Event time series 2 (one means an event, or zero otherwise)}

\item{tau}{Parameter for filtering function (See fmetric function from mmpp
package.)}

\item{repetitions}{Number of repetitions to construct the confidence interval}

\item{sig_level}{The significance level to test if correlation is significant.}
}
\value{
0 if significant synchronization or 1 otherwise
}
\description{
This function compares the times which the events occur e.g., time indices
where the time series values are different than zero. Note that the intensity
does not matter but if there is an event or not. This function also performs
a statistical test using a shuffling approach to test significance. This
implementation uses the fmetric from the mmpp package.
}
