% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_net.R
\name{tsnet_qn}
\alias{tsnet_qn}
\title{Construct a quantile network from a time series.}
\usage{
tsnet_qn(x, breaks, weighs_as_prob = TRUE, remove_loops = FALSE, ...)
}
\arguments{
\item{x}{Array. Time series}

\item{breaks}{The breaks used to find the intervals. It can be either
an integer defining the number of breaks (automatically found) or
an array of real values defining the breaks. To check the intervals
used, see the nodes' `range` attribute using \link[igraph]{get.vertex.attribute}
  See \link[base]{cut} function for more details.}

\item{weighs_as_prob}{Boolean. If TRUE (default), return the transition
probabilities as in a markov chain. Otherwise, returns the counts
each transition appears.}

\item{remove_loops}{Boolean. If TRUE, remove loops.}

\item{...}{Additional parameters to \link[base]{cut} function.}
}
\value{
quantile network
}
\description{
Construct a quantile network from a time series.
}
