% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_cities_temperature_df}
\alias{us_cities_temperature_df}
\title{Monthly temperatures in 27 US cities.}
\format{
A data frame with 61 rows and 28 variables:
\describe{
  \item{date}{First day of each month. The temperatures correspond to the mean values of each month }
  \item{other columns}{Temperature (ºC) time series in each city}
}
}
\source{
\url{https://www.kaggle.com/datasets/selfishgene/historical-hourly-weather-data}
}
\usage{
us_cities_temperature_df
}
\description{
A data set containing the temperature in 27 US cities from 2012 and 2017. This data set
was adapted from the original and considers only cities in the US. Data was grouped by
month (mean value) and removed days with missing data.
}
\note{
This data set is released under the Database Contents License (DbCL) v1.0.
}
\author{
David Beniaguev
}
\keyword{dataset}
