% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_cities_temperature_list}
\alias{us_cities_temperature_list}
\title{Monthly temperatures in 27 US cities.}
\format{
A list with a time series for each city. For the dates, check us_cities_temperature_df
  data set.
}
\source{
\url{https://www.kaggle.com/datasets/selfishgene/historical-hourly-weather-data}
}
\usage{
us_cities_temperature_list
}
\description{
A data set containing the temperature in 27 US cities from 2012 and 2017. This data set
was adapted from the original and considers only cities in the US. Data was grouped by
month (mean value) and removed days with missing data.
}
\note{
This data set is released under the Database Contents License (DbCL) v1.0.
}
\author{
David Beniaguev
}
\keyword{dataset}
