% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_matrix_normalize}
\alias{dist_matrix_normalize}
\title{Normalize a distance/similarity matrix.}
\usage{
dist_matrix_normalize(D, to = c(0, 1))
}
\arguments{
\item{D}{Distance/similarity matrix}

\item{to}{An array of two elements c(min_value, max_value) representing
the interval where the elements of dist_matrix will be normalized to.}
}
\value{
Normalized matrix
}
\description{
Normalize a distance/similarity matrix.
}
