% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_net.R
\name{net_weighted}
\alias{net_weighted}
\title{Creates a weighted network.}
\usage{
net_weighted(D, invert_dist_as_weight = TRUE)
}
\arguments{
\item{D}{Distance matrix. All values must be between [0,1].}

\item{max_dist_value.}{The value}
}
\value{
Fully connected network
}
\description{
A link is created for each pair of nodes, except if the distance is
maximum (1). In network science, stronger links are commonly represented
by high values. For this reason, the link weights returned are 1 - D.
}
