% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_cor}
\alias{tsdist_cor}
\title{Positive or negative correlation distance.}
\usage{
tsdist_cor(ts1, ts2, positive_cor = TRUE)
}
\arguments{
\item{ts1}{Array. Time series 1}

\item{ts2}{Array. Time series 2}

\item{positive_cor}{Boolean. If TRUE (default), only positive correlations are considered.
If FALSE, only negative correlations are considered.}
}
\value{
Real value [0,1] where 0 means perfect positive (or negative
if positive_cor==F) correlation and 1 no positive (or negative
if positive_cor==F) correlation.
}
\description{
Perfect positive returns zero and one means no  or negative correlations. The
opposite occurs if positive_cor==F.
}
