% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_cor}
\alias{tsdist_cor}
\title{Absolute, positive, or negative correlation distance.}
\usage{
tsdist_cor(ts1, ts2, cor_type = "abs")
}
\arguments{
\item{ts1}{Array. Time series 1}

\item{ts2}{Array. Time series 2}

\item{cor_type}{String. "abs" (default), "+", or "-". "abs" considers the
correlation absolute value. "+" only positive correlations and "-" only
negative correlations.}
}
\value{
Real value [0,1] where 0 means perfect positive (or negative
if positive_cor==F) correlation and 1 no positive (or negative
if positive_cor==F) correlation.
}
\description{
Perfect positive returns zero and one means no  or negative correlations. The
opposite occurs if positive_cor==F.
}
