% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dist.R
\name{tsdist_eca}
\alias{tsdist_eca}
\title{Event coincidence analysis.}
\usage{
tsdist_eca(ets1, ets2, delT, tau, sym, ...)
}
\arguments{
\item{ets1}{Event time series 1 (one means an event, or zero otherwise)}

\item{ets2}{Event time series 2 (one means an event, or zero otherwise)}

\item{...}{TODO: FINISH!!!}
}
\value{

}
\description{
This function compares the times which the events occur e.g., time indices
where the time series values are different than zero. Note that the intensity
does not matter but if there is an event or not. This function also performs
a statistical test using a shuffling approach to test significance. This
implementation uses the CC.eca.ts function from the CoinCalc package.
}
