% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_net.R
\name{tsnet_vg}
\alias{tsnet_vg}
\title{Construct the visibility graph from a time series}
\usage{
tsnet_vg(x, method = c("nvg", "hvg"), limit = +Inf, num_cores = 1)
}
\arguments{
\item{x}{Array. Time series}

\item{method}{String. Construction method: "nvg" (default) for
Natural visibility graph, "hvg" horizontal visibility graph.}

\item{limit}{Positive integer. The maximum temporal distance (indexes)
allowed in the visibility. This parameter limits the max visibility.}

\item{num_cores}{Number of cores (default = 1).}
}
\value{
visibility graph
}
\description{
TODO: weights
}
